Add libtest.h and libtest.c along with basic functions
libtest.h contains the declaration of ct_init, ct_report, and
ct_terminate, which constitute the interface for the ctest library.

ct_init Reports the name of the module to be tested, and initializes
testing statistics.

ct_report Prints the name of one test, and records the test status.

ct_terminate Prints a summary of test statistics and frees memory.